+++ test/scratchspace.jl
@@ -1,0 +1,18 @@
+
+# Scratch space for testing development features
+
+using PowerPlots
+using PowerModels
+using PowerPlots, PowerModels
+data = parse_file("$(joinpath(dirname(pathof(PowerModels)), ".."))/test/data/matpower/case5.m")
+
+################################################
+## No code should be commited in this section ##
+################################################
+
+
+
+
+
+
+################################################